+++ ui/app/src/views/ViewDashboard.tsx
@@ -19,6 +19,7 @@ import { PluginRegistry } from '@perses-dev/plugin-system';
 import { bundledPluginLoader } from '../model/bundled-plugins';
 import { useDashboard } from '../model/dashboard-client';
 import { useDatasourceApi } from '../model/datasource-api';
+import DashboardBreadcrumbs from '../components/DashboardBreadcrumbs';
 
 /**
  * The View for viewing a Dashboard.
@@ -49,7 +50,13 @@ function ViewDashboard() {
       <ErrorBoundary FallbackComponent={ErrorAlert}>
         <PluginRegistry pluginLoader={bundledPluginLoader}>
           <ErrorBoundary FallbackComponent={ErrorAlert}>
-            <DashboardView dashboardResource={data} datasourceApi={datasourceApi} />;
+              <DashboardView
+                dashboardResource={data}
+                datasourceApi={datasourceApi}
+                dashboardTitleComponent={
+                  <DashboardBreadcrumbs dashboardName={data.metadata.name} dashboardProject={data.metadata.project} />
+                }
+              />
           </ErrorBoundary>
         </PluginRegistry>
       </ErrorBoundary>